requirements.txt
asgiref==3.7.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
Django==4.2.2
djangorestframework==3.14.0
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.2
whitenoise==6.4.0
